refactor(browser): extract watch runtime lifecycle
