+++ packages/browser/src/hostController.ts
@@ -5,7 +5,6 @@ import {
   buildBrowserCoverageMap,
   type ExecutorCycleOutcome,
   type ExecutorInvalidationCallback,
-  FATAL_SIGNALS,
   type ListBrowserTestsOptions,
   color,
   createCoverageProvider,
@@ -94,6 +93,7 @@ import {
   type SourceMapPayload,
 } from './sourceMap/sourceMapLoader';
 import { collectWatchTestFiles, planWatchRerun } from './watchRerunPlanner';
+import { registerWatchCleanup, watchContext } from './watchRuntime';
 
 /**
  * Monotonic counter for synthetic per-file Perfetto `pid` values in `--trace`
@@ -356,30 +356,6 @@ export class ContainerRpcManager {
 }
 
 // ============================================================================
-// Browser Runtime - Core runtime state
-// ============================================================================
-
-// ============================================================================
-// Watch Mode Context - Process-lifecycle watch state
-// ============================================================================
-
-// Only process-wide concerns stay module-level: the runtime handle reused
-// across controller re-entry (config-change restarts), and the signal/exit
-// cleanup that must run once per process. Diff/rerun state lives on
-// `BrowserRuntime.watchState`.
-type WatchContext = {
-  runtime: BrowserRuntime | null;
-  cleanupRegistered: boolean;
-  cleanupPromise: Promise<void> | null;
-};
-
-const watchContext: WatchContext = {
-  runtime: null,
-  cleanupRegistered: false,
-  cleanupPromise: null,
-};
-
-// ============================================================================
 // Utility Functions
 // ============================================================================
 
@@ -444,71 +420,6 @@ const resolveProviderForTestPath = ({
     `Cannot resolve browser provider for test path: ${JSON.stringify(testPath)}. ` +
       `Known project roots: ${JSON.stringify(sortedProjects.map((p) => p.rootPath))}`,
   );
-};
-
-/**
- * Tear down the persistent watch runtime (dev servers, provider, browser,
- * WebSocket server). Idempotent, and the single teardown the browser executor's
- * `close` and the process-exit nets both go through.
- */
-export const runWatchRuntimeTeardown = <T>(
-  state: { runtime: T | null; cleanupPromise: Promise<void> | null },
-  destroy: (runtime: T) => Promise<void>,
-): Promise<void> => {
-  if (state.cleanupPromise) {
-    return state.cleanupPromise;
-  }
-
-  state.cleanupPromise = (async () => {
-    if (!state.runtime) {
-      return;
-    }
-
-    await destroy(state.runtime);
-    state.runtime = null;
-  })();
-
-  // The memo is released once this teardown settles, because the state outlives
-  // the session: a config-file change restarts the run against a fresh runtime,
-  // and a memo left resolved from the previous session would make every later
-  // teardown a no-op — leaving the session after that to reuse a runtime built
-  // from the pre-restart config. Idempotency only has to hold within a runtime.
-  return state.cleanupPromise.finally(() => {
-    state.cleanupPromise = null;
-  });
-};
-
-export const cleanupWatchRuntime = (): Promise<void> =>
-  // `cleanupRegistered` is deliberately not re-armed alongside the memo: the
-  // signal nets it installs read `watchContext.runtime` live, so they stay
-  // correct across a restart, and re-registering would stack a fresh set of
-  // listeners on every one.
-  runWatchRuntimeTeardown(watchContext, destroyBrowserRuntime);
-
-const registerWatchCleanup = (embedded: boolean): void => {
-  if (watchContext.cleanupRegistered) {
-    return;
-  }
-  watchContext.cleanupRegistered = true;
-
-  // Embedded (programmatic) hosts own the process lifecycle; they tear the
-  // session down through the browser executor's `close` instead of signals.
-  if (embedded) {
-    return;
-  }
-
-  // Cleanup-only nets: core's watch loop owns the signal → exit-code path and
-  // awaits the same idempotent `cleanupWatchRuntime` promise through the
-  // browser executor's `close`.
-  for (const signal of FATAL_SIGNALS) {
-    process.once(signal, () => {
-      void cleanupWatchRuntime();
-    });
-  }
-
-  process.once('exit', () => {
-    void cleanupWatchRuntime();
-  });
 };
 
 // ============================================================================
